+++ app/components/first_draft_collections/update/form_component.html.erb
@@ -7,5 +7,5 @@
     html: { class: 'needs-validation collection-editor', novalidate: true, multipart: true } do |form| %>
   <%= render FirstDraftCollections::Create::DetailsComponent.new(form: form) %>
   <%= render Collections::SettingsComponent.new(form: form) %>
-  <%= render FirstDraftCollections::Create::ButtonsComponent.new(form: form) %>
+  <%= render FirstDraftCollections::ButtonsComponent.new(form: form) %>
 <% end %>
